requirements.txt
requests~=2.26.0
python-dotenv~=0.20.0
terminaltables~=3.1.10